fix: updated integration hubstaff plugin by NX generator
