+++ apps/gauzy/src/app/pages/hubstaff/components/hubstaff/hubstaff.component.scss
@@ -1,23 +1,26 @@
+@import 'gauzy/_gauzy-table.scss';
+@import 'gauzy/_gauzy-cards.scss';
+
 .hubstaff-container {
-	min-height: 200px;
+  min-height: 200px;
 }
 
 .card-header {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
 }
 
 ::ng-deep {
-	.hubstaff-container {
-		.angular2-smart-actions {
-			text-align: center;
-			width: 5%;
+  .hubstaff-container {
+    .angular2-smart-actions {
+      text-align: center;
+      width: 5%;
 
-			.form-control {
-				width: 15px;
-				display: inline-block;
-			}
-		}
-	}
+      .form-control {
+        width: 15px;
+        display: inline-block;
+      }
+    }
+  }
 }
