+++ packages/core/package.json
@@ -51,7 +51,6 @@
 		"@gauzy/contracts": "^0.1.0",
 		"@gauzy/integration-ai": "^0.1.0",
 		"@gauzy/integration-github": "^0.1.0",
-		"@gauzy/integration-hubstaff": "^0.1.0",
 		"@gauzy/integration-jira": "^0.1.0",
 		"@gauzy/plugin": "^0.1.0",
 		"@godaddy/terminus": "^4.12.1",
