+++ packages/core/src/integration-entity-setting-tied/index.ts
@@ -1,0 +1,3 @@
+export * from './integration-entity-setting-tied.module';
+export * from './integration-entity-setting-tied.service';
+export * from './integration-entity-setting-tied';
