+++ packages/core/src/integration-entity-setting-tied/integration-entity-setting-tied.seed.ts
@@ -1,18 +1,16 @@
 import { DataSource } from 'typeorm';
 import { faker } from '@faker-js/faker';
 import { IIntegrationEntitySettingTied, IntegrationEntity, ITenant } from '@gauzy/contracts';
-import { PROJECT_TIED_ENTITIES } from '@gauzy/integration-hubstaff';
 import { IntegrationEntitySetting, IntegrationTenant, Organization } from './../core/entities/internal';
 import { IntegrationEntitySettingTied } from './integration-entity-setting-tied.entity';
+import { PROJECT_TIED_ENTITIES } from './integration-entity-setting-tied';
 
 export const createRandomIntegrationEntitySettingTied = async (
 	dataSource: DataSource,
 	tenants: ITenant[]
 ): Promise<IIntegrationEntitySettingTied[]> => {
 	if (!tenants) {
-		console.warn(
-			'Warning: tenants not found, Integration Entity Setting  will not be created'
-		);
+		console.warn('Warning: tenants not found, Integration Entity Setting  will not be created');
 		return;
 	}
 
@@ -35,26 +33,18 @@ export const createRandomIntegrationEntitySettingTied = async (
 
 				integrationEntitySettingTiedEntity.integrationEntitySetting = integrationEntitySetting;
 				integrationEntitySettingTiedEntity.sync = faker.datatype.boolean();
-				integrationEntitySettingTiedEntity.organization = faker.helpers.arrayElement(
-					organizations
-				);
+				integrationEntitySettingTiedEntity.organization = faker.helpers.arrayElement(organizations);
 				integrationEntitySettingTiedEntity.tenant = tenant;
 				//todo: need to understand real values here
-				if (
-					integrationEntitySetting['entity'] ===
-					IntegrationEntity.PROJECT
-				) {
-					integrationEntitySettingTiedEntity.entity = faker.helpers.arrayElement(
-						PROJECT_TIED_ENTITIES
-					)['entity'];
+				if (integrationEntitySetting['entity'] === IntegrationEntity.PROJECT) {
+					integrationEntitySettingTiedEntity.entity =
+						faker.helpers.arrayElement(PROJECT_TIED_ENTITIES)['entity'];
 				} else {
 					integrationEntitySettingTiedEntity.entity = faker.helpers.arrayElement(
 						Object.values(IntegrationEntity)
 					);
 				}
-				randomIntegrationEntitySettingsTiedEntity.push(
-					integrationEntitySettingTiedEntity
-				);
+				randomIntegrationEntitySettingsTiedEntity.push(integrationEntitySettingTiedEntity);
 			}
 		}
 	}
