+++ packages/core/src/integration-entity-setting/index.ts
@@ -1,0 +1,3 @@
+export * from './integration-entity-setting.module';
+export * from './integration-entity-setting.service';
+export * from './integration-entity-settings';
