+++ packages/core/src/integration-map/index.ts
@@ -1,3 +1,3 @@
 export * from './integration-map.module';
 export * from './integration-map.service';
-export { IntegrationMapSyncEntityCommand } from './commands';
+export * from './commands';
