+++ packages/core/src/integration/index.ts
@@ -1,0 +1,2 @@
+export * from './integration.module';
+export * from './integration.service';
