--- packages/plugins/integration-hubstaff/.dockerignore
@@ -1,10 +1,0 @@
-.git
-.gitignore
-.gitmodules
-README.md
-docker
-node_modules
-tmp
-build
-dist
-.env
