+++ packages/plugins/integration-hubstaff/README.md
@@ -1,7 +1,11 @@
-# Integration HubStaff
+# Integration Hubstaff
 
 This library was generated with [Nx](https://nx.dev).
 
+## Building
+
+Run `nx build integration-hubstaff` to build the library.
+
 ## Running unit tests
 
-Run `ng test integration-hubstaff` to execute the unit tests via [Jest](https://jestjs.io).
+Run `nx test integration-hubstaff` to execute the unit tests via [Jest](https://jestjs.io).
