--- packages/plugins/integration-hubstaff/jest.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-	name: 'integration-hubstaff',
-	preset: '../../../jest.config.js',
-	transform: {
-		'^.+\\.[tj]sx?$': 'ts-jest'
-	},
-	moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
-	coverageDirectory: '../../coverage/plugins/integration-hubstaff'
-};
