+++ packages/plugins/integration-hubstaff/package.json
@@ -1,7 +1,7 @@
 {
 	"name": "@gauzy/integration-hubstaff",
 	"version": "0.1.0",
-	"description": "Ever Gauzy Platform plugin for integration with HubStaff APIs",
+	"description": "Ever Gauzy Platform plugin for integration with Hubstaff APIs",
 	"author": {
 		"name": "Ever Co. LTD",
 		"email": "ever@ever.co",
@@ -9,30 +9,64 @@
 	},
 	"license": "AGPL-3.0",
 	"private": true,
-	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
+	"type": "commonjs",
+	"main": "./dist/src/index.js",
+	"types": "./dist/src/index.d.ts",
 	"files": [
 		"dist"
 	],
 	"directories": {
-		"lib": "dist",
-		"test": "__test__"
+		"lib": "dist"
 	},
 	"publishConfig": {
 		"access": "restricted"
 	},
-	"scripts": {
-		"test:e2e": "jest --config ./jest.config.js",
-		"build": "rimraf dist && yarn run compile",
-		"compile": "tsc -p tsconfig.lib.json"
+	"peerDependencies": {
+		"tslib": "^2.6.2"
 	},
-	"keywords": [],
 	"dependencies": {
-		"@gauzy/contracts": "^0.1.0"
+		"@gauzy/common": "^0.1.0",
+		"@gauzy/config": "^0.1.0",
+		"@gauzy/contracts": "^0.1.0",
+		"@gauzy/core": "^0.1.0",
+		"@gauzy/plugin": "^0.1.0",
+		"@nestjs/axios": "^3.0.2",
+		"@nestjs/common": "^10.3.7",
+		"@nestjs/cqrs": "^10.2.7",
+		"@nestjs/swagger": "^7.3.0",
+		"axios": "^1.6.8",
+		"typeorm": "^0.3.20",
+		"moment": "^2.30.1",
+		"rxjs": "^7.4.0"
 	},
 	"devDependencies": {
 		"@types/node": "^17.0.33",
 		"rimraf": "^3.0.2",
 		"typescript": "5.1.6"
-	}
+	},
+	"keywords": [
+		"Hubstaff",
+		"integration",
+		"time tracking",
+		"productivity",
+		"API",
+		"plugin",
+		"NestJS",
+		"Gauzy",
+		"Ever Gauzy",
+		"platform",
+		"management",
+		"tool",
+		"software",
+		"development",
+		"typescript"
+	],
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/ever-co/ever-gauzy"
+	},
+	"bugs": {
+		"url": "https://github.com/ever-co/ever-gauzy/issues"
+	},
+	"homepage": "https://ever.co"
 }
