+++ packages/plugins/integration-hubstaff/project.json
@@ -1,0 +1,48 @@
+{
+	"name": "integration-hubstaff",
+	"$schema": "../../../node_modules/nx/schemas/project-schema.json",
+	"sourceRoot": "packages/plugins/integration-hubstaff/src",
+	"projectType": "library",
+	"targets": {
+		"build": {
+			"executor": "@nrwl/js:tsc",
+			"outputs": ["{options.outputPath}"],
+			"options": {
+				"outputPath": "./packages/plugins/integration-hubstaff/dist",
+				"tsConfig": "packages/plugins/integration-hubstaff/tsconfig.lib.json",
+				"packageJson": "packages/plugins/integration-hubstaff/package.json",
+				"main": "packages/plugins/integration-hubstaff/src/index.ts",
+				"assets": ["packages/plugins/integration-hubstaff/*.md"]
+			}
+		},
+		"publish": {
+			"executor": "nx:run-commands",
+			"options": {
+				"command": "node tools/scripts/publish.mjs integration-hubstaff {args.ver} {args.tag}"
+			},
+			"dependsOn": ["build"]
+		},
+		"lint": {
+			"executor": "@nrwl/linter:eslint",
+			"outputs": ["{options.outputFile}"],
+			"options": {
+				"lintFilePatterns": ["packages/plugins/integration-hubstaff/**/*.ts"]
+			}
+		},
+		"test": {
+			"executor": "@nrwl/jest:jest",
+			"outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+			"options": {
+				"jestConfig": "packages/plugins/integration-hubstaff/jest.config.ts",
+				"passWithNoTests": true
+			},
+			"configurations": {
+				"ci": {
+					"ci": true,
+					"codeCoverage": true
+				}
+			}
+		}
+	},
+	"tags": []
+}
