+++ packages/plugins/integration-hubstaff/src/lib/hubstaff.config.ts
