+++ packages/plugins/integration-hubstaff/src/lib/integration-hubstaff.plugin.ts
@@ -1,0 +1,30 @@
+import { GauzyCorePlugin as Plugin, IOnPluginBootstrap, IOnPluginDestroy } from '@gauzy/plugin';
+import { ApplicationPluginConfig } from '@gauzy/common';
+import { HubstaffModule } from './hubstaff.module';
+
+@Plugin({
+	imports: [HubstaffModule],
+	configuration: (config: ApplicationPluginConfig) => config
+})
+export class IntegrationHubstaffPlugin implements IOnPluginBootstrap, IOnPluginDestroy {
+	// We disable by default additional logging for each event to avoid cluttering the logs
+	private logEnabled = true;
+
+	/**
+	 * Called when the plugin is being initialized.
+	 */
+	onPluginBootstrap(): void | Promise<void> {
+		if (this.logEnabled) {
+			console.log(`${IntegrationHubstaffPlugin.name} is being bootstrapped...`);
+		}
+	}
+
+	/**
+	 * Called when the plugin is being destroyed.
+	 */
+	onPluginDestroy(): void | Promise<void> {
+		if (this.logEnabled) {
+			console.log(`${IntegrationHubstaffPlugin.name} is being destroyed...`);
+		}
+	}
+}
