--- packages/plugins/integration-hubstaff/tslint.json
@@ -1,4 +1,0 @@
-{
-	"extends": "../../../tslint.json",
-	"rules": []
-}
